add noindent parameter to all boxes
When SILE indentation is enabled all boxes are indented by 20pt from
frame left. Calling `\noindent` before `\framebox` doesn't solve this;
the only way I have managed to fix it is to call `\neverindent`, which
is undesirable if you want indentation elsewhere.

This adds a new parameter `noindent` to all box types, which shifts the
boxes left by `20pt` when `true`.

It seems to have to be hard-coded to `20pt` - getting
`current.parindent` doesn't seem to work. Would be nice to figure out
why, but I'm at a loss.